refactor: align file names and placement with responsibility
Naming (name now matches contents):
- lib/subagent.ts → lib/session-key.ts (session-key helpers, not just subagent)
- services/agent/schema.ts → types.ts (held one interface, no zod schema)
- lib/glob.ts → lib/glob-match.ts (event-name matching, not FS globbing)

Placement (file now lives with its responsibility):
- Merge channels/contracts.ts into channels/types.ts; drop the
  backward-compat re-export shim and duplicate type imports.
- lib/heartbeat.ts → services/cron/ and lib/skills.ts → services/agent/
  (single-consumer helpers).
- lib/error-store.ts → services/agent/ (agent-only domain state).
- channels/media-extract.ts → media-paths.ts (disambiguate from
  media/extract-document).
- media/extract-document.ts → media/providers/document.ts (sits with the
  other format handlers).

Tests moved/renamed to follow their subjects. Pure moves + import-path
edits; no behavior change.
