+++ services/channels/providers/whatsapp/adapter.ts
@@ -6,8 +6,7 @@ import { readFileSync, existsSync, writeFileSync } from 'node:fs';
 import path from 'node:path';
 import { jidDecode, jidNormalizedUser, areJidsSameUser } from '@whiskeysockets/baileys';
 import type { WASocket } from '@whiskeysockets/baileys';
-import type { InboundMediaSource } from '../../types.js';
-import type { NormalizedInboundMessage, AdapterDeps } from '../../contracts.js';
+import type { InboundMediaSource, NormalizedInboundMessage, AdapterDeps } from '../../types.js';
 import { BaseChannelAdapter, MEDIA_TYPE_LABELS } from '../../base-adapter.js';
 import { createWhatsAppSocket } from './session.js';
 import type { WhatsAppInboundMessage } from './types.js';
